+++ src/rex.deploy/src/htsql_rex_deploy/__init__.py
@@ -1,15 +1,21 @@
 #
-# Copyright (c) 2013, Prometheus Research, LLC
+# Copyright (c) 2014, Prometheus Research, LLC
 #
 
 
 from htsql.core.addon import Addon
-from . import classify
+from . import classify, introspect, tr
 
 
 class DeployAddon(Addon):
 
     name = 'rex_deploy'
+    hint = """configure rex.deploy databases"""
+    help = """
+    This addon provides metadata configuration for database
+    schemas generated with `rex.deploy`.
+    """
+    packages = ['.', '.tr']
 
     @classmethod
     def get_extension(cls, app, attributes):