+++ src/rex.deploy/src/htsql_rex_deploy/tr/__init__.py
@@ -1,0 +1,8 @@
+#
+# Copyright (c) 2014, Prometheus Research, LLC
+#
+
+
+from . import lookup
+
+