+++ src/rex.deploy/NEWS.rst
@@ -5,6 +5,17 @@
 .. contents:: Table of Contents
 
 
+1.2.0 (2014-XX-XX)
+==================
+
+* Store metadata as comments on tables, columns and other entities.
+* Preserve and restore table, column and link labels when the SQL name is
+  mangled.
+* Added table, column and link titles.
+* Added HTSQL plugin that generates HTSQL configuration from ``rex.deploy``
+  metadata.
+
+
 1.1.0 (2013-12-20)
 ==================